// src/FX.d.ts
declare enum EffectSequenceComponentType {
    Sprite = 0,
    MovieClip = 1,
    Emitter = 2,
    Trigger = 3
}
// ...
export interface IBaseEffect {
    name: string;
    id: any;
    type: number;
    containerId: string;
}
// ...
export interface IEffectSequenceSettings extends IBaseEffect {
    effects: IEffectSettings[];
    delay: number;
    scaleMin: number;
    scaleMax: number;
}
export interface IEffectSettings {
    componentId: string;
    componentType: EffectSequenceComponentType;
    delay: number;
    componentParams: IMovieClipComponentParams | IBaseComponentParams;
    scaleMin: number;
    scaleMax: number;
    alphaMin: number;
    alphaMax: number;
    rotationMin: number;
    rotationMax: number;
    blendMode: number;
    tint: number;
    duration: number;
    containerId: string;
    triggerValue: string;
}
// ...
export interface IMovieClipComponentParams extends IBaseComponentParams {
    animationSpeedMin: number;
    animationSpeedMax: number;
    loop: boolean;
}
export interface IBaseComponentParams {
    anchorX: number;
    anchorY: number;
}